Create GsonPayloadAdapter to encrypt/decrypt WalletConnect payloads
